% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_weather.R
\docType{data}
\name{klimaFR}
\alias{klimaFR}
\title{Monthly climate in Freiburg}
\format{A data frame with 732 observations on 8 variables:
\describe{
  \item{year}{begin of interval (year)}
  \item{month}{begin of interval (month)}
  \item{temp_mean}{monthly mean of daily temperature means in 2m height in °C}
  \item{temp_max}{monthly mean of daily temperature maxima in 2m height in °C}
  \item{temp_min}{monthly mean of daily temperature minima in 2m height in °C}
  \item{wind_speed}{monthly mean of daily wind speed in Bft}
  \item{sun_dur}{monthly sum of sunshine duration in h}
  \item{precip}{monthly sum of precipitation height in mm}
}}
\source{
DWD Climate Data Center (CDC): Historical monthly station observations (temperature, pressure, precipitation, sunshine duration, etc.) for Germany, version v007, 2018. (\url{http://dwd.de})
}
\usage{
klimaFR
}
\description{
This data set contains various informations about the climate in Freiburg
    from 1950 to 2010 in a monthly rhythm.
}
\examples{
## mean of the max. temperatures
mean(klimaFR$temp_max, na.rm = TRUE)
## sum of sunshine duration in 1950
sum(klimaFR$sun_dur[1:12])
## smallest an largest min. temperature
min(klimaFR$temp_min, na.rm = TRUE)
max(klimaFR$temp_min, na.rm = TRUE)
## precipitation in every december
klimaFR$precip[klimaFR$month == "12"]
}
\keyword{datasets}
