% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_weather.R
\name{plot_weather}
\alias{plot_weather}
\title{Lineplot of the weather}
\usage{
plot_weather(x, ...)

## with extra arguments
plot_weather(x, yr)
plot_weather(x, yr, both = FALSE)
}
\arguments{
\item{x}{a column from klimaFR. This data frame will set the weather condition.}

\item{...}{further arguments passed to or from other methods}

\item{yr}{a integer between 1950 and 2010 or a numeric array with integers between 1950 and 2010. The chosen years will be used to generate the third and fourth plot.}

\item{both}{a boolean. This boolean will decide whether you will get both grids or only the grid with the chosen years in yr.}
}
\value{
The described grid
}
\description{
Generic function to create one grid which includes two or four lineplots.
    The first two plots show the chosen weather condition during the years from 1950 to 2010 and during 1950, 1980 and 2010.
    The last two show the chosen weather condition during the years from the earliest to the latest of the chosen years and during the chosen years.
    If the parameter both is false the first two plots will not be visible.
}
\note{
Without the packages ggplot2, scales, cowplot and reshape2 this function won't work.
If you want to have a better plot (e.g. without overlapping titles), use the zoom (in RStudio).
Please keep in mind that the plots might have gaps because of the missing infomation in the data frame.
}
\examples{
plot_weather(klimaFR$temp_mean)

plot_weather(klimaFR$precip, yr =  c(2000, 1999, 2006))

plot_weather(klimaFR$sun_dur, yr =  c(1965, 1954, 1960), both = FALSE)

}
