% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowtotext.R
\name{rowtotext}
\alias{rowtotext}
\title{Print a row}
\usage{
rowtotext(x)
}
\arguments{
\item{x}{a date (string)}
}
\value{
The described text
}
\description{
Creates a text which contains all information about the weather conditions in Freiburg
    in the chosen month and year.
    Please keep in mind that the year has to be between 1950 and 2010.
}
\note{
The format of the date should follow the format \%b \%Y
    (see: \url{https://www.statmethods.net/input/dates.html})
}
\examples{
rowtotext("Jan 1967")
rowtotext("Aug 2002")
##next example contains at least one NA
rowtotext("Oct 2002")
}
